Ignore issue about undefined variable
Using &$error inside the use clause of a Closure to make a variable
defined inside the closure available outside should be fine. Psalm seems
to be OK with it on the playground, which uses the master branch, which
must contain a fix, but I did not find which commit exactly on master
fixes it.